fix(command & commnds): fixed When updating amount of one product without amount
The products like 'Pesca Esportiva' and 'Pesque Pague' are special products because they don't have
amount, and when we try to update the amount of product in backend we got null. The problem it was
right after this the code was updating in frontend the product, so how it was null, we got typerror.
I've just made an if/else to see if it's null or not in stockProductsReducer of Commands and Command
page